BE-Refactor: reservation 모델 인덱스 추가, 타입별 조회 status 쿼리 옵션 추가
